refactor: move utils to eval_utils/
